requirements.txt
fastapi
uvicorn
requests
python-multipart
httpx
python-dotenv
Pillow
rembg